Version 3.3.1
IMPROVEMENTS: This version focusses on the Task Validation function, therefore on two attrbutes 'duedate' and 'description'. The LLM has its own current date which dates back to 2023, hence that needed to be overwritted (line 630 app.py), also a new validation for duedate was added which gave the status: 'ON TRACK', 'OVERDUE', 'DUE TODAY' or 'DUE DATE MISSING' (The working is yet to be tested completely. NOTE: Validation of deadline using description is giving errors or hallucinting due to the absence of Description